+++ Logic/EditItem.cpp
@@ -29,18 +29,18 @@ public:
 	}
 
 	void executeAction(vector<Item> &vectorStore) {
-		if(_lineNumber == 0   || _lineNumber > vectorStore.size()) {
+		if(_lineNumber == 0   || _lineNumber > (int)vectorStore.size()) {
 			_message = ERROR_INVALID_LINE_NUMBER + " ";
-			_message += ('0'+_lineNumber);
+			_message += ('0' + _lineNumber);
 			_message += "\n";
 		} else {
 			char buffer[1000];
 			sprintf_s(buffer, SUCCESS_EDITED.c_str(), 
-				vectorStore[_lineNumber-1].toString().c_str(), 
+				vectorStore[_lineNumber - 1].toString().c_str(), 
 				_input.toString().c_str());
 			_message=buffer;
 
-			vectorStore[_lineNumber-1] = _input;
+			vectorStore[_lineNumber - 1] = _input;
 		}
 	}
